refactor(desktop): split oversized entry-actions and intent-actions hooks
The repo-shape check flags both files over the 400-line soft threshold
after the review fix-up. Extract the workspace-entry selection deps and
the pending-interaction resolution actions into sibling hooks; public
APIs unchanged.
